visa3/api/requirements.txt
fastapi
uvicorn
mysql-connector-python
pydantic
requests
python-jose[cryptography]
passlib[bcrypt]
python-multipart
